:art: Moved multi-cluster app to wiregaurd-operator
