+++ operators/wireguard/apps/multi-cluster/Containerfile
