+++ operators/wireguard/apps/multi-cluster/apps/client/main.go
@@ -1,8 +1,8 @@
 package client
 
 import (
-	"github.com/kloudlite/operator/apps/multi-cluster/apps/client/env"
-	"github.com/kloudlite/operator/apps/multi-cluster/mpkg/wg"
+	"github.com/kloudlite/operator/operators/wireguard/apps/multi-cluster/apps/client/env"
+	"github.com/kloudlite/operator/operators/wireguard/apps/multi-cluster/mpkg/wg"
 	"github.com/kloudlite/operator/pkg/logging"
 )
