+++ operators/wireguard/apps/multi-cluster/apps/server/main.go
@@ -2,8 +2,8 @@ package server
 
 import (
 	"github.com/gofiber/fiber/v2"
-	"github.com/kloudlite/operator/apps/multi-cluster/apps/server/env"
-	"github.com/kloudlite/operator/apps/multi-cluster/mpkg/wg"
+	"github.com/kloudlite/operator/operators/wireguard/apps/multi-cluster/apps/server/env"
+	"github.com/kloudlite/operator/operators/wireguard/apps/multi-cluster/mpkg/wg"
 	"github.com/kloudlite/operator/pkg/logging"
 )
