+++ operators/wireguard/apps/multi-cluster/apps/server/server.go
@@ -5,9 +5,9 @@ import (
 	"sync"
 
 	"github.com/gofiber/fiber/v2"
-	"github.com/kloudlite/operator/apps/multi-cluster/apps/common"
-	"github.com/kloudlite/operator/apps/multi-cluster/apps/server/env"
-	"github.com/kloudlite/operator/apps/multi-cluster/mpkg/wg"
+	"github.com/kloudlite/operator/operators/wireguard/apps/multi-cluster/apps/common"
+	"github.com/kloudlite/operator/operators/wireguard/apps/multi-cluster/apps/server/env"
+	"github.com/kloudlite/operator/operators/wireguard/apps/multi-cluster/mpkg/wg"
 	"github.com/kloudlite/operator/pkg/logging"
 )
