+++ operators/wireguard/apps/multi-cluster/helm/multi-cluster/templates/_namespace.yml
