+++ operators/wireguard/apps/multi-cluster/helm/multi-cluster/templates/agents.yml
