+++ operators/wireguard/apps/multi-cluster/mpkg/text-template/main.go
