+++ operators/wireguard/apps/multi-cluster/templates/wg-client.conf.tpl
