+++ operators/wireguard/apps/multi-cluster/templates/wg-server.conf.tpl
