--- operators/wireguard/apps/wg-restart/.gitignore
@@ -1,1 +1,0 @@
-conf.json
